mythtv: add ffmpeg-cutter user job script
This is an earlier attempt than keyframe-cutter that solely relies on
ffmpeg for cutting and merging. Unfortunately ffmpeg only accepts
timestamps for start/stop locations.

For this we need to convert the MythTV cut list data, which is in
keyframe number, to time stamps. During testing it turned out that that
conversion is too inaccurate, i.e. we either cut out too late or cut in
too soon, so that in the new, merged recording there are still
left-overs from the stuff we want to cut out from the original
recording.