mythtv: add watchdog server script
Ensure that mediabox really wakes up when it is supposed to:

- ACPI wakeup timer failed to wake the system
- power loss and BIOS wakeup on powerloss didn't work
- when this server script is started

This script needs to be installed on a permanently running system. It
listens for watchdog updates from mediabox. It schedules a timeout one
minute after the desired wakeup time. It uses etherwake to send a WOL
event to the mediabox when the timeout expires.