mythtv: add kernel patches for si2168 driver
MythTV DVBv5 support is currently borked/incomplete and doesn't allow
to correctly set DVB-C mode on the Hauppauge WinTV-dualHD DVB-C/T/T2
USB receiver. The patch adds an option to the si2168 driver to disable
DVB-T/T2 completely, so that the receiver automatically runs in DVB-C
mode. Unfortunately the patch got rejected by the kernel developers and
hence we have to rebuild the patched driver on every kernel update :-(

The patches are based on Linux kernel v5.4.

Assuming the patched dvb-frontends/ drivers source directory under the
current directory the following steps are required to build & install
the patched kernel module:

 # modprobe -r em28xx_rc em28xx_dvb em28xx si2168 si2157 && lsmod | fgrep -e si21 -e dvb -e em28
 # make -C dvb-frontends/ clean
 # make -C dvb-frontends/ modules
 # mkdir -p /lib/modules/$(uname -r)/kernel/extras && cp dvb-frontends/si2168.ko /lib/modules/$(uname -r)/kernel/extras && restorecon -vR /lib/modules/$(uname -r)/kernel/extras && depmod -ae -F /boot/System.map-$(uname -r) $(uname -r) && modinfo si2168
 # modprobe em28xx_rc em28xx_dvb em28xx si2168 si2157 && lsmod | fgrep -e si21 -e dvb -e em28