multimedia: add DPMS toggler
HDMI audio only works when the HDMI output is active. For that reason we
need to disable DPMS when we want to output audio via HDMI to a receiver
for a longer period of time.

This small script uses kdialog to generate a dialog that offers the
current DPMS status and a switch to toggle it.