multimedia: add DBus screensaver inhibitor
Under Wayland we can't use DPMS anymore. Instead use DBus to inhibit the
screensaver. Unfortunately we need to keep the DBus connection open for
the inhibit to stay active. Add a helper tool that goes to sleep after
inhibiting the screensaver. It is up to the caller to send a kill to the
process to uninhibit the screensaver again.