mythtv: add shutdown script
The wrapper is called by MythTV to shut down the system. The wrapper
calls the real script with sudo as we need to execute root commands in
it. Currently we simply shutdown machine, relying on the RTC CMOS wake
alarm or the watchdog to start the system at a later time.