mythtv: add shutdown check & alarm script
The wrapper is called by MythTV to check if it is OK to shutdown the
system (without any parameters) or when the system should wake up the
next time (date & time as parameters).

The wrapper calls the real script with sudo as we might have need root
rights to access system files.

The script writes the wake up time to the RTC CMOS wake alarm and the
watchdog server. It checks if we need to wake up earlier to run the EPG
update. Sometimes MythTV doesn't seem to set the wake up time. For this
case we always record the last time and use it during shutdown check to
update the RTC CMOS & watchdog.