comics: add page break between comics
As it turns out the HTML parser accepts a H1 element as a page break. Append a
H1 element to every node we insert. Convert the dummy page at the end to normal
text as there is now already a page break before it.

Now we get a sensible Loc counter when reading the document on a Kindle.