multimedia: update for ALSA HDMI output changes
After the latest Fedora 37 updates the number of HDMI outputs and their
numbering has changed in ALSA and thus in PulseAudio/Pipewire. Update the ouput
name in the script.

NOTE: device 8 on the Nvidia card has disappeared. The correct device for the
hdmi-playback PCM alias definition in .asoundrc is now 3 (aplay -l: HDMI 0).